Refactor: split collect.py into collectors/ package
Each collector lives in its own file under collectors/. Shared helpers
(_utcnow, _iso, _default_since, _strip_html) are in collectors/_helpers.py.
Dispatch logic (PLATFORM_COLLECTORS, collect_all) is in collectors/_dispatch.py.
collect.py becomes a backward-compat shim.

--platform choices in run.py are now generated from PLATFORM_COLLECTORS.keys()
rather than a hardcoded list, with tests to verify this.
