+++ collectors/_types.py
@@ -1,0 +1,3 @@
+from __future__ import annotations
+from typing import Any
+CollectorResult = dict[str, Any] | None
